requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
openapi-spec-validator>=0.7.1
openapi-core>=0.22.0
pyyaml>=6.0.1
httpx>=0.27.0
sentence-transformers>=3.0.1
pydantic>=1.10.0,<3
python-dotenv>=1.0.1
jinja2>=3.1.4
python-multipart>=0.0.9
